requirements.txt
Flask==2.1.2
python-dotenv==0.21.0
Werkzeug==2.1.2
WTForms==3.0.1
